// src/lib/api/models/DetailedBookableItemDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DetailedBookableItemCategoryDto } from './DetailedBookableItemCategoryDto';
export type DetailedBookableItemDto = {
    id?: number;
    name?: string | null;
    description?: string | null;
    category?: DetailedBookableItemCategoryDto;
};